backend/services/web/requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
flask-cors
